[domaindescriptions] move PolygonalDomain to domaindescriptions.polygonal
move DiscDomain, CircularSectorDomain as well
